Scrapper/requirements.txt
googlemaps==4.5.3
itemadapter==0.4.0
itemloaders==1.0.4
lorem==0.1.1
pandas==1.3.3
peewee==3.14.4
python-dotenv==0.19.1
Scrapy==2.5.0
scrapy_selenium==0.0.7
tqdm==4.62.3
